Simplify precision-specification in tests.
Now that assertEquals considers dtypes in determining tolerance, most
tests don't need explicitly set precision.

Those that do are a few half precision tests on cuda. In this PR, those
are broken out to be handled explicitly, though we may also want to
consider further loosening the tolerance on half-precision.

ghstack-source-id: 642a428f65ab0014fcb0d99b1efb1aa9485e5307
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37181